translation/requirements.txt
fairseq
torch
moses
subword_nmt
sacremoses
pycloud-client==0.7.1
